Created a repository for tea-break, an app to remind users to stop and go have some coffee every hour.
The initial plan is to do a GUI in Java, consisting of a window with a
message ("time to take a break"), and two buttons to dismiss the window, or
display a randomly generated "reason" to take a break. Regular invocation
of the program will be handled by cron, which allows us to have the program
run every 60 minutes after the computer has booted.