Created a new class for the dismiss button, which simply exits the app when it is clicked.
Also modified the main window class so that closing the window via the
window controls will also exit the app.

Note that because the periodic scheduling is done by cron, the app will
start up again an hour after it is closed (once the cron scripting work is
in place at least). I am considering options for ways to switch off the
alarm entirely; one possibility would be to use the existence of a
particular file (eg. /var/run/tea-break/disable). Another alternative
would be to have some kind of config file, but that is more involved, as
we would have to parse it.